pyproject.toml
[project]
name = "luna"
version = "0.1.0"
description = "Luna - Personal Telegram Assistant with Google Calendar integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "anthropic>=0.76.0",
    "apscheduler>=3.11.2",
    "google-api-python-client>=2.188.0",
    "google-auth-oauthlib>=1.2.4",
    "mcp>=1.26.0",
    "pytelegrambotapi>=4.30.0",
    "python-dotenv>=1.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/luna"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""
